Use dynamic inventory with libvirt
Use community.libvirt inventory plugin to talk to guests. This should
just work with this commit:

    $ ansible all -i inventory.yml -m ansible.builtin.ping

This needs a running QEMU guest agent in the guest with the right RPC
capabilities set. Also note that we set remote_tmp = /tmp in ansible.cfg
for the guest agent to properly function. SELinux labels do not allow
the guest agent to write to /root/.ansible/tmp directory, the default
path.

References:
- https://docs.ansible.com/ansible/latest/plugins/inventory.html